utils: add a root broker
This will allow core to drop the sudo requirement and will be a
much-smaller code to inspect for security issues.

When time permits, it should be re-written in Rust, to further
restrict the capabilities.

A C client library will also be needed for env_dump to dump some
debugfs files.