smartezbench: introduce the DONE state
Mark the state of a profile as DONE when it is done enhancing the
report. Reset the state to RUN if adding more tests.

This will also have the added benefit of reducing the start time of
ezbenchd as it will not have to check all the old reports to check
if there is more work to be done.