utils/owatch: Overwatch, a watchdog wrapper
Overwatch will execute a program given as its parameters, quite like
`timeout'. It will then monitor the stdout and stderr output of the
child, and any activity on either will act as a heartbeat. If there is
no heartbeat in the given timeout period, the child process (group)
will be killed.

If Overwatch is run as root, it will also use a hardware watchdog
(/dev/watchdog[0-9]*) if it can open one. It will signal the hardware
watchdog on child process heartbeat and will just stop pinging it
(thus causing a reboot) if the child process does not heartbeat.

v2: (Martin Peres)
 - trailing space deletion
 - add the copyrights

Signed-off-by: Petri Latvala <petri.latvala@intel.com>